@@ -13,28 +13,28 @@
     "@types/classnames": "^2.2.10",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-helmet": "^5.0.8",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
+    "@types/react-helmet": "^6.1.11",
     "axios": "^0.19.0",
     "classnames": "^2.2.6",
     "fs-extra": "^9.0.1",
     "mobx": "^5.11.0",
     "mobx-react-lite": "^1.4.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-helmet": "^5.2.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-scripts": "5.0.1",
     "react-swipeable": "^5.4.0",
     "rimraf": "^3.0.2",
-    "typescript": "^4.0.3",
+    "typescript": "^5.9.3",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "NODE_OPTIONS=--openssl-legacy-provider react-scripts start",
+    "build": "NODE_OPTIONS=--openssl-legacy-provider react-scripts build",
+    "test": "NODE_OPTIONS=--openssl-legacy-provider react-scripts test",
+    "eject": "NODE_OPTIONS=--openssl-legacy-provider react-scripts eject",
     "deploy": "node deploy.js"
   },
   "homepage": "https://dreadboy.github.io/giraffe",
