@@ -1,13 +1,16 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import {Root} from './components/Root';
 import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
+ReactDOM
+    .createRoot(
+        document.getElementById('root') as HTMLElement
+    )
+    .render(
     <React.StrictMode>
         <Root/>
-    </React.StrictMode>,
-    document.getElementById('root'),
+    </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
